requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
docopt==0.6.2
idna==3.3
pipreqs==0.4.11
python-dotenv==0.19.2
requests==2.27.1
slack-bolt==1.11.4
slack-sdk==3.15.1
urllib3==1.26.8
yarg==0.1.9
